+++ concordium-std/src/impls.rs
@@ -1,4 +1,4 @@
-use crate::{convert, mem, num, prims::*, traits::*, types::*};
+use crate::{convert, mem, num, prims, prims::*, traits::*, types::*};
 use concordium_contracts_common::*;
 
 use mem::MaybeUninit;
@@ -452,7 +452,7 @@ impl HasLogger for Logger {
     }
 
     #[inline(always)]
-    fn log_bytes(&mut self, event: &[u8]) {
+    fn log_raw(&mut self, event: &[u8]) {
         unsafe {
             log_event(event.as_ptr(), event.len() as u32);
         }
@@ -478,10 +478,15 @@ impl HasActions for Action {
     }
 
     #[inline(always)]
-    fn send(ca: &ContractAddress, receive_name: &str, amount: Amount, parameter: &[u8]) -> Self {
-        let receive_bytes = receive_name.as_bytes();
+    fn send_raw(
+        ca: &ContractAddress,
+        receive_name: ReceiveName,
+        amount: Amount,
+        parameter: &[u8],
+    ) -> Self {
+        let receive_bytes = receive_name.get_chain_name().as_bytes();
         let res = unsafe {
-            send(
+            prims::send(
                 ca.index,
                 ca.subindex,
                 receive_bytes.as_ptr(),
@@ -527,6 +532,22 @@ pub fn put_in_memory(input: &[u8]) -> *mut u8 {
     #[cfg(not(feature = "std"))]
     core::mem::forget(bytes);
     ptr
+}
+
+/// Wrapper for
+/// [HasActions::send_raw](./trait.HasActions.html#tymethod.send_raw), which
+/// automatically serializes the parameter. Note that if the parameter is
+/// already a byte array or convertible to a byte array without allocations it
+/// is preferrable to use [send_raw](./trait.HasActions.html#tymethod.send_raw).
+/// It is more efficient and avoids memory allocations.
+pub fn send<A: HasActions, P: Serial>(
+    ca: &ContractAddress,
+    receive_name: ReceiveName,
+    amount: Amount,
+    parameter: &P,
+) -> A {
+    let param_bytes = to_bytes(parameter);
+    A::send_raw(ca, receive_name, amount, &param_bytes)
 }
 
 impl<A, E> UnwrapAbort for Result<A, E> {