+++ concordium-std/CHANGELOG.md
@@ -6,6 +6,12 @@
   to be consistent with the Write implementation for ContractState.
 - Use little-endian encoding for sender contract addresses in receive contexts. This
   reverts the change in concordium-std 0.4.1.
+- Change the `receive_name` parameter of `HasActions::send` to use `ReceiveName`
+  instead of `str`.
+- Rename `send` to `send_raw` in `HasActions`.
+- Rename `log_bytes` to `log_raw` in `HasLogger`.
+- Add `send`, a wrapper for `HasActions::send_raw`, which automatically
+  serializes `parameter` (using `Serial`).
 - Allow init and receive methods to return custom error codes that will be displayed to the user
   if a smart-contract invocation fails.
 - Add i128 and u128 support to serialization and schema.